Move `FilterByDate` into API concerns
To be consistent with other concerns; making the include statements simpler.
