+++ app/controllers/api/v2/participants_controller.rb
@@ -2,7 +2,7 @@ module API
   module V2
     class ParticipantsController < BaseController
       include Pagination
-      include ::API::Concerns::FilterByDate
+      include FilterByDate
 
       def index
         render json: to_json(paginate(participants_query.participants))
