+++ app/controllers/api/v3/declarations_controller.rb
@@ -2,7 +2,7 @@ module API
   module V3
     class DeclarationsController < BaseController
       include Pagination
-      include ::API::Concerns::FilterByDate
+      include FilterByDate
 
       def index
         render json: to_json(paginate(declarations_query.declarations))
