+++ app/controllers/api/v3/statements_controller.rb
@@ -2,7 +2,7 @@ module API
   module V3
     class StatementsController < BaseController
       include Pagination
-      include ::API::Concerns::FilterByDate
+      include FilterByDate
 
       def index
         render json: to_json(paginate(statements_query.statements))
